% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_clustering.R
\name{edge_clustering}
\alias{edge_clustering}
\title{Edge communities and contribution of edges}
\usage{
edge_clustering(links, vec_col = c("#aa380b", "#0b58aa", "#f7e37e"))
}
\arguments{
\item{links}{set of links.}

\item{vec_col}{a vector describing three colors.}
}
\description{
Figure 3. Network diagram showing edge communities and contribution of edges from each community to node themes.
}
\author{
Helen Wheeler, Kevin Cazelles
}
