% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{links}
\alias{links}
\title{Links as found in fig 1 in Wheeler et al. (2018)}
\format{A data frame of 42 rows and 14 columns characterizing relatoinships
between monitoring items.}
\usage{
data(links)
}
\description{
Links as found in fig 1 in Wheeler et al. (2018)
}
\author{
Helen Wheeler
}
\keyword{datasets}
