% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{links2}
\alias{links2}
\title{Links as found in fig 2a and b in Wheeler et al. (2018)}
\format{A data frame.}
\usage{
data(links2)
}
\description{
Links as found in fig 2a and b in Wheeler et al. (2018)
}
\author{
Helen Wheeler
}
\keyword{datasets}
