% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{links3}
\alias{links3}
\title{Links as found in fig 2c in Wheeler et al. (2018)}
\format{A data frame.}
\usage{
data(links3)
}
\description{
Links as found in fig 2c in Wheeler et al. (2018)
}
\author{
Helen Wheeler
}
\keyword{datasets}
