% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcomp2a.R
\name{netcomp2a}
\alias{netcomp2a}
\title{Network of desired impacts from monitoring}
\usage{
netcomp2a(links, nodes, vec_col1 = c("#f1bf5f", "#00ac89", "#27ade3"),
  vec_col2 = c("#d72968", "#a18037", "#5523b4", "#063632", 1))
}
\arguments{
\item{links}{set of links.}

\item{nodes}{set of nodes.}

\item{vec_col1}{a vector describing three colors.}

\item{vec_col2}{a vector describing five colors.}
}
\description{
Fig S1: Plotting of network of desired impacts from monitoring -
comparison of three algorithms.
}
\author{
Helen Wheeler, Kevin Cazelles
}
