% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nodes}
\alias{nodes}
\title{Nodes as found in fig 1 in Wheeler et al. (2018)}
\format{A data frame of 18 rows and 7 columns describing the montoring items
as in Wheeler et al. (2018).}
\usage{
data(nodes)
}
\description{
Nodes as found in fig 1 in Wheeler et al. (2018)
}
\author{
Helen Wheeler
}
\keyword{datasets}
