% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nodes2}
\alias{nodes2}
\title{Nodes as found in fig 2a and b in Wheeler et al. (2018)}
\format{A data frame.}
\usage{
data(nodes2)
}
\description{
Nodes as found in fig 2a and b in Wheeler et al. (2018)
}
\author{
Helen Wheeler
}
\keyword{datasets}
