% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nodes3}
\alias{nodes3}
\title{Nodes as found in fig 2c in Wheeler et al. (2018)}
\format{A data frame.}
\usage{
data(nodes3)
}
\description{
Nodes as found in fig 2c in Wheeler et al. (2018)
}
\author{
Helen Wheeler
}
\keyword{datasets}
