% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walktrap_dendro.R
\name{walktrap_dendro}
\alias{walktrap_dendro}
\title{Community detection, and network and dendrogram plotting using
the walktrap algorithm}
\usage{
walktrap_dendro(links, nodes, vec_col1 = c("#edac35", "#64c4eb", "#00ac89"),
  vec_col2 = c("#d72968", "#a18037", "#5523b4", "#063632", 1))
}
\arguments{
\item{links}{set of links.}

\item{nodes}{set of nodes.}

\item{vec_col1}{a vector describing three colors.}

\item{vec_col2}{a vector describing five colors.}
}
\description{
Fig 2 a,b: Community detection, and network and dendrogram plotting using
the walktrap algorithm.
}
\author{
Kevin Cazelles, Helen Wheeler
}
