% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity2a.R
\name{connectivity2a}
\alias{connectivity2a}
\title{Cumulative weights of intra- and inter-group connections for each node}
\usage{
connectivity2a(links, nodes, vec_col1 = c("#f1bf5f", "#00ac89",
  "#27ade3"))
}
\arguments{
\item{links}{set of links.}

\item{nodes}{set of nodes.}

\item{vec_col1}{a vector describing three colors.}
}
\description{
Figure 2c. Cumulative weights of intra- and inter-group
connections for each node.
}
\author{
Kevin Cazelles, Helen Wheeler
}
