% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dartdiag.R
\name{dartdiag}
\alias{dartdiag}
\title{dart-styled diagram: Network x closness}
\usage{
dartdiag(nodes, links, vec_col = c("#edac35", "#64c4eb", "#00ac89"))
}
\arguments{
\item{nodes}{set of nodes.}

\item{links}{set of links.}

\item{vec_col}{a vector describing three colors.}
}
\description{
Figure 1. Network diagram of the desirable impacts of monitoring
as identified by 29 stakeholders in arctic monitoring.
}
\author{
Kevin Cazelles, Helen Wheeler
}
